requirements.txt
daily-python
requests
python-dotenv
azure-cognitiveservices-speech
azure-cognitiveservices-speech==1.38.0
Flask==3.0.3
Flask-SocketIO==5.3.6
pydub==0.25.1
requests==2.32.3